Add explicit field IDs to flatbuffer schemas (#148)
### Changelog
Added explicit field IDs to flatbuffer schemas.

### Docs

None

### Description

This allows the schemas to be compiled with `--require-explicit-ids`
turned on in `flatc`, which can be nice to help prevent people from
accidentally making breaking changes to flatbuffer schemas if they
choose to reorder fields (or insert new fields in the middle of the
schemas). Because of how flatc works with includes, if downstream users
want to use `--require-explicit-ids` on their own schemas that happen to
include foxglove schema types, they must update the foxglove schemas
that they use to include IDs.