rust: Rework websocket blocking interface (#191)
I realized that we really need to expose a first-class blocking
interface for the websocket server, and not just for python. We really
don't want async-clueless programmers slapping a `#[tokio::main]` on
their entrypoint and then blocking tokio threads by accident. That'll be
a nightmare to detect and debug.

In this change:
- Add `WebSocketServer::tokio_runtime` to bring your own runtime
(unstable).
- Add `ws-demo-blocking.rs` example to demo a program without
`#[tokio::main]`.
- Split `WebSocketServerHandle` into a blocking and non-blocking
version.