Turn rust examples into packages (#199)
This change re-works the rust examples. Instead of all the examples
being sources as part of the "foxglove" package each example is now its
own package within the `rust/examples` folder. This makes it more
obvious which dependencies are needed for an example and isolates the
examples from each other. A reader can "copy-paste" an example as a
complete starting point (including the Cargo.toml file).

This change also does away with the "examples/common" folder. I found
this layer of indirection confusing for understanding the code I would
need to write.

Co-authored-by: Greg Smith <gasmith@foxglove.dev>