Add rust to generate script (#209)
### Changelog
None

### Docs
None

### Description
Extended `yarn generate` script to call `cargo run --bin
foxglove-proto-gen`, required to generate the rust code.

The rust generation code cannot be directly ported to typescript because
it relies on a rust library (prost) to generate its protobuf
implementation.

The only remaining codegen that does not live in `yarn generate` is for
python; this will become obsolete once
https://github.com/foxglove/foxglove-sdk/pull/201 lands.