Generate pyo3 schemas for SDK (#201)
This PR produces well-known Foxglove schemas as `pyclass`es which can be
used by [pyo3](https://pyo3.rs/) in the SDK. For the default channel
logging interface, we plan to remove protobuf serialization on the
python side, and instead use the existing serialization in SDK core.
We've demonstrated that this can be done with little relative overhead;
see FG-10358.

Most of the changes here are in `generatePyclass.ts`; per FG-10347,
we're going to simplify build steps between schemas and SDK. For now,
this produces output in a `pyclass` directory, and the build script for
the python SDK copies those files into the project. (Open to suggestions
here.)

The generator produces two files:

- the schemas, expressed as `pyclass`-annotated Rust structs & enums,
- the .pyi stub interface for the above,

Currently, the submodule is registered with a function call instead of a
declarative module. This mechanism is set to be deprecated in pyo3, but
integrating with hand-written python isn't straightforward. Once we have
the build process unified from FG-10347, it'll be a little easier to
make the declarative submodule syntax work.