Python SDK: integrate new channels and schemas (#210)
This PR integrates the new generated Schema and typed Channel classes
from #201 and #206 respectively, and updates the examples accordingly.

As part of this, I've adjusted the Schema classes to make them easier to
work with:
- fields are now `Optional`, with defaults. Callers can omit any
properties during construction.
- constructors are keyword-only, which should be more future-proof.

The protobuf dependency is no longer needed in the Python SDK.

For discussion: I've also re-worked the high-level API a bit. It works
with any given foxglove Schema now (though could use better typing). But
it no longer works with arbitrary protobuf messages. We could support
that, but I'm not sure how valuable it is. We could support arbitrarily
serialized data, but then you'd need to provide your schema, and the
channel interface seems much nicer.