Update MCAP handling in Python SDK (#219)
This fixes the implementation of `PyMcapWriter` to agree with the
declared interface. Notably, it exposes the `close` method and documents
the auto-closing behavior if a reference to the writer is not held.

To make this easier to work with, this also adds a context-managed
function `new_mcap_file`, which frees the user from needing to maintain
a reference to the writer. I included 'mcap' in the name since (a) we've
decided to expose that name in the Rust SDK, and (b) it makes clear with
the output format will be, since we're asking the user to supply a file
name.

This also updates the example to produce a basic MCAP file.