Update Python example and remove explicit encoders (#216)
This updates the Python example and underlying module to better
represent where we're headed.

The logging API serves the three use cases below.

1. Custom schemas and serialization.
2. Custom schemas, where the SDK serializes for you. 
3. Messages well-known Foxglove schemas.

This PR adds an example for use case (1), using JSON. I think we could
add an example using protobuf, but I don't want to add protobuf again as
a dev dependency. We could consider organizing examples like we've done
with Rust — each getting its own package.

Previously, the examples made use of provided `Encoder`s. This sort-of
addresses use case (2), but for JSON still required users to provide a
JSON schema. We plan to address this in a better way (FG-10466).

Use case (2) is already represented by the examples, and is unchanged
here.

However, for discussion: I've also kept it easy to work with JSON —
particularly for exploratory use cases. You can pass schemas and
messages as dicts, and these will be treated as JSON.