python: improve documentation and auto-generate api docs (#223)
This improves the documentation of native types, and adds the
`sphinx.ext.autodoc` extension to auto-generate Python module
documentation.

We're currently using Sphinx v7 as v8 requires a newer version of
Python, and v7 is not able to generate docs from the stub files.

The docs are not yet deployed anywhere but can be built from the
python SDK directory using the documented command.