rust: Add support for websocket services (#213)
This change adds support for hosting custom services on the websocket
server. As far as I can tell, this is mostly intended to support the
ros-foxglove-bridge.

This turned into a pretty large change, but I'm not sure that splitting
it up into advertisement vs. call and response (as the linear tickets
are arranged) would've made things any easier.

The basic ideas:
 - Add the `Service` type (id, name, schema, handler).
 - Store an <Id, Service> map on the websocket server.
 - Provision a buffered response channel for each client.
 - Advertise and unadvertise services as necessary.
 - Handle calls:
   - Allocate a response (or reject with "too many requests")
   - Lookup service by ID (or reject with "unknown service")
   - Invoke handler.

The handler itself is a flexible trait, so that it can be as simple as a
`Fn` closure, or as complex as you like. At the moment, the handler runs
synchronously from the context of the client poll loop, but you can
always `tokio::spawn` your way to a less-constrained environment.

While writing tests (and the perhaps-overblown example), I realized that
I really _really_ wanted a full-featured crate for round-tripping
ws-protocol messages. I think we might want to pull
`websocket::protocol` out into its own `foxglove-ws-protocol` crate. Any
objections?

And because it's a bit buried, you can check out the service
registration demo
[here](https://github.com/foxglove/foxglove-sdk/pull/213/files#diff-2dc63a716bbefb10ba159933f2a1d3efec9caa0c5cde4bac7b39073ec726d81bR24-R72).

Fixes: FG-9672
Fixes: FG-9674

---------

Co-authored-by: Eloff <dan.eloff@gmail.com>