ci: Merge release tags with --ff when available (#225)
### Changelog
None

### Docs
None

### Description
When merging release tags, there is no need for `--no-ff`. The thing we
want to avoid is a squash merge (which would alter the commit we
tagged). But if there have been no new commits to main since the release
is tagged, then `--ff` is fine.

By removing `--no-ff`, git will automatically create either a
fast-forward merge or a merge commit, as needed.