rust: Support the rest of parameter protocol (#212)
This fills out support for the remainder of the parameter protocol.

I tried to follow the behavior of the Python server, with the exception
of the bug I uncovered, which is fixed here. on_parameters_subscribe and
on_parameters_unsubscribe is only called when a parameter is first
subscribed to by a client, or last unsubscribed to. This corresponds to
when you first need to worry about it, and when you can stop worrying
about it, without having to do that bookkeeping yourself.

Something I copied, but am unsure of, is how set_parameters will not
update parameters if there's no listener installed for the server.
Should that be the behavior?

Marked as a draft because I still need to add tests.